pyproject.toml
[tool.poetry]
name = "amahr-ds561-hw2"
version = "0.1.0"
description = ""
authors = ["Alicja Mahr <bmahr@bu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
networkx = "^3.1"
numpy = "^1.26.0"
tqdm = "^4.66.1"
scipy = "^1.11.2"
google-cloud-storage = "^2.11.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
